+++ src/hack_flexispot/mod.rs
@@ -1,0 +1,2 @@
+pub mod command;
+pub mod device;
